* 'RubikPi-c6490p-le1.0-dev-r1.0_00075.0' of ssh://10.0.80.23:29418/general/kernel/qcom: (3139 commits)
  Revert "BACKPORT: UPSTREAM: arm64: dts: qcom: sa8775p: add CPU idle states"
  Revert "FROMLIST: arm64: dts: qcom: sa8775p: Add CPUs to psci power domain"
  BACKPORT: arm64: dts: qcom: sc7280: don't enable GPU on unsupported devices
  QCLINUX: drm/msm/a6xx: Fix UBWC configuration
  PENDING: misc: fastrpc: Remove PD status check
  QCLINUX: arm64: deconfig: enable CONFIG_FW_DEVLINK_SYNC_STATE_TIMEOUT
  PENDING: tty: serial: qcom-geni-serial: fix CTS control for UART in loopback mode
  QCLINUX: arm64: dts: qcom: qcs9075-addons: Rename rb8 to iq-9075-evk
  QCLINUX: dt-bindings: arm: qcom: qcs9075-addons: Rename rb8 to iq-9075-evk
  PENDING: arm64: dts: qcom: Rename qcs9075-rb8 to qcs9075-iq-9075-evk
  PENDING: dt-bindings: arm: qcom: Rename qcs9075-rb8 to qcs9075-iq-9075-evk
  PENDING: qcom: ice: Add ice probe
  PENDING: ASoC: qcom: qcm9100: Fix the dapm routings for qcs8300
  QCLINUX: ASoC: qcom: qcm6490: Use qcom_snd_parse_of_v2 to support common dailink data
  PENDING: arm64: dts: qcom: qcs9075: Change phy-mode to SGMII
  PENDING: net: stmmac: Fix NULL pointer access for GPTP packets
  QCLINUX: arm64: defconfig: enable Novatek NT36672E and Lontium lt9611UXC configs
  UPSTREAM: arm64: dts: qcom: qcs6490-rb3gen2: enable hdmi bridge
  UPSTREAM: arm64: defconfig: enable Novatek NT36672E DSI Panel driver
  UPSTREAM: arm64: dts: qcom: qcm6490-idp: add display and panel
  ...

+++ drivers/dma-buf/heaps/qcom_system_heap.c
@@ -37,7 +37,7 @@
  *	Andrew F. Davis <afd@ti.com>
  *
  * Copyright (c) 2020-2021, The Linux Foundation. All rights reserved.
- * Copyright (c) 2022-2023 Qualcomm Innovation Center, Inc. All rights reserved.
+ * Copyright (c) 2022-2024 Qualcomm Innovation Center, Inc. All rights reserved.
  */
 
 #include <linux/dma-buf.h>
@@ -494,8 +494,8 @@ free_mem:
 
 static struct dma_buf *system_heap_allocate(struct dma_heap *heap,
 					    unsigned long len,
-					    unsigned long fd_flags,
-					    unsigned long heap_flags)
+					    u32 fd_flags,
+					    u64 heap_flags)
 {
 	struct qcom_sg_buffer *buffer;
 	DEFINE_DMA_BUF_EXPORT_INFO(exp_info);
